Clarify Godep vs GOPATH in documentation
The immediate impetus for these changes was that
the `restore` command had unclear help text.
Clarifying it required some matching changes
elsewhere.

Note that the verb "check out" in `restore`'s
Long usage text is also used in a similar context
in `go help get`:

"When checking out or updating a package, get
looks for a branch or tag that matches the locally
installed version of Go."

Fixes #33.