remove old code just before copying in new code
It's relatively likely for an error to occur while
loading the dependency list for the given packages. (For
example, if the user doesn't have the go tool in their
search path.) In this case, 'godep save' would delete
the entire Godeps tree before exiting with an error
message.

Now we wait to remove the existing tree until just
before copying in the code. We also do as much as
possible, in particular writing out the new mainfest,
before deleting the old tree.

Fixes #53.