save: new flag -r to rewrite import paths
One drawback of godep's previous default behavior is
that repos using it cease to be 'go gettable'. This is a
big problem because it impacts a lot of people: not just
the repo's developers, but also its users.

Flag -r solves this by rewriting import paths to refer
directly to the copied source code at its new location.
So a package C that depends on package D will actually
import C/Godeps/_workspace/src/D. This makes C's repo
self-contained and causes 'go get' to build C with the
right version of all dependencies.

fixes #68