save: test more implementation corner cases
Add two more tests in preparation for #99.

Check that a dependency and transitive dependency are
correctly copied in place from separate gopath roots. If
dependency D is in root1/src/D, and T is in root2/src/T,
and D depends on T, make sure that both D and T can be
found and copied into the current package.

Also check that we properly exclude subdirectories of
dependencies even when the subdirectory is obtained as a
rewritten import path from another dependency. If we
depend on D and T, and D depends on T/X (rewritten as
D/Godeps/_workspace/src/T/X), list only T in our
manifest, not T/X.