Pass --force when checking out git repos
When a bare git repo has already been checked out and it is checked out
again into a different directory, only files changed between HEAD and
the new commit will appear in the new directory. The rest of the files
will be treated as unstaged deletions.

Passing --force changes this to discard local changes (i.e. the
perceived deletion of files) and checks out all the files.