fix migrating from -copy=false to true
Flag -copy is used to determine where to write out the
newly-generated manifest, but it was also incorrectly
being used to locate any existing manifest file. This
meant that when switching to or from -copy, reading the
existing manifest was broken. We fix this by attempting
to read an existing manifest in both locations,
regardless of the given flags.

We also explicitly check that the correct version of
each dependency package is installed in GOPATH, when
switching to -copy=true.