Handle imports for ignored go files
Analyze the imports and build tags of IgnoredGoFiles and resolve those
dependencies adding them back into the originating package’s imports.
This also goes part of the way to eliminating `go list` as it has to do
at least part of the way to eliminating go list as similar code can be
used to resolve imports / tags for all go files.

Also introduced a package cache for LoadPackages, this is also used to
short circuit recursive calls to LoadPackages