restore: download all in 1 pass, then restore
Modify restore so that it downloads all the dependencies in one pass,
and then runs restore for each dependency in a second pass. We have seen
an issue where, depending on the order of the `go get` downloads,
sometimes one or more dependencies end up with the master branch as
HEAD, instead of the revision specified in Godeps.json. It appears that
`go get` resets the HEAD to master, even if the dependency has already
been downloaded and is pointing to a different ref.