run command once during restore -v
Running `godep restore -v` would always error out because the
CombinedOutput command did a Run under the hood, and the second Run at
the end of restore would error out.

Along the way, I added some more context (and a couple of newlines) to
the restore error prints to figure out what was going on.