In restore errors, mention which package had the problem.
This improves the situation when the user runs `godep restore`
and sees a few instances of "restore, during download dep" errors.
Previously, it was not mentioned which packages failed to download.
Turning on -v verbose mode was not a good solution since it prints
too much other information for successfully processed packages.