Rewrite packages only inside the project
When calling "godep -r ./... github.com/example/proj", godep will
import the example/proj and its dependencies into our project tree and
will rewrite the paths accordingly. But it will also modify the
example/proj repository in GOPATH. In other words - the example/proj
modified in two places - in source directory (in
$GOPATH/src/github.com/example/proj) and in destination directory (in
./Godeps/_workspace/src/github.com/example/proj).

This commit fixes it, so the packages outside the cwd are not
rewritten.