Replace `go list`
Main motivation is so that we can attempt to pick up packages across
architectures and to have more control over the list of packages.

I consider this the first in several structural changes.

Most additions are inside of list.go.

For each run of listPackage a depScanner is made to track already seen
packages (inspired by Scanner). That list is not cached between
executions.

The current implementation of listPackage attempts to resolve vendor/ a
bit better than we used to, but other changes need to be implemented
elsewhere in godep before that can be fully realized.

All *existing* tests pass, but this doesn't mean it actually works
*better*.

A test was added to show that we fix #271.