Parse files ourselves
I tried to use build.Context to handle most of the grunt work, but that
didn't work out that well.

So now we use it to help us find the package and then we use
ast.ParseFile to parse the *.go files in a package and then fill in the
rest of the build.Package

We ignore all `// +build ignore` tagged files.

Fixes #348