handle detached git on restore
If the git repo is detached, we need to determine the default remote
branch before attempting to donwload (can't pull into a repo in detached
state).

This attempts to use "git remote show origin" to have the origin tell us
what the default branch is.

PS: this may not work with old (~3 year old git servers).